helm/requirements.txt
# Tooling for keeping the helm-chart up-to-date
semver
frigate
toml
pyyaml
